backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
pydantic
pydantic-settings
passlib
bcrypt
python-jose
face_recognition
opencv-python
numpy
